% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseStataBin.R
\name{chooseStataBin}
\alias{chooseStataBin}
\title{Set Stata binary path}
\usage{
chooseStataBin()
}
\description{
Set Stata binary (among found alternatives) path. These settings are
lost when R is closed, therefore you should consider adding a
\code{options("RStata.StataPath")} line in your \code{.Rprofile}.
}

