% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStata-package.R
\docType{package}
\name{RStata}
\alias{RStata}
\alias{RStata-package}
\title{A Bit of Glue Between R and Stata}
\description{
A simple R -> Stata interface allowing the user to execute Stata
commands (both inline and from a .do file) from R.
}

