% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_stata_config.R
\name{set_stata_config}
\alias{set_stata_config}
\title{set_stata_config}
\usage{

  set_stata_config(StataPath = "/Applications/Stata/StataSE.app/Contents/MacOS/StataSE",
  StataVersion = 15)
}
\arguments{
\item{StataPath}{local path to stata configuraion}

\item{StataVersion}{integer denoting the version of stata}
}
\value{
Printed statement
}
\description{
Point R session to local copy of Stata.
}
\examples{
\dontrun{
set_stata_config(StataPath = '/Applications/Stata/StataSE.app/Contents/MacOS/StataSE',StataVersion = 15)
}
}
