% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stata}
\alias{stata}
\title{Send commands to a Stata process}
\usage{
stata(x)
}
\arguments{
\item{x}{character vector of length 1 (path to \code{.do} file) or more
(a set of stata commands). See examples.}
}
\description{
Simple functions to send commands to a Stata process
}
\examples{
## Single command
stata("help regress")
## Many commands
do <- c( "set obs 200",
         "gen a = 1"   )
stata(do)
## External file
stata("foo.do")
}

